changed id to long in Bookings (Caused issues with int)

+++ src/main/java/com/example/bookingfriseur/main/entities/Bookings.java
@@ -24,7 +24,7 @@ public class Bookings {
         return barber;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
@@ -35,7 +35,7 @@ public class Bookings {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
 
-    private int id;
+    private long id;
 
     private String first_name;
     private String last_name;
